+++ src/specs/NativeCastModule.ts
@@ -1,30 +1,30 @@
 /* eslint-disable @typescript-eslint/no-wrapper-object-types */
 import type { TurboModule } from 'react-native';
 import { TurboModuleRegistry } from 'react-native';
-import { Int32 } from 'react-native/Libraries/Types/CodegenTypes';
+import { Double } from 'react-native/Libraries/Types/CodegenTypes';
 
 export interface Spec extends TurboModule {
-  casting(tag: Int32): Promise<boolean>;
+  casting(tag: Double): Promise<boolean>;
 
-  chromecastCasting(tag: Int32): Promise<boolean>;
+  chromecastCasting(tag: Double): Promise<boolean>;
 
-  airplayCasting(tag: Int32): Promise<boolean>;
+  airplayCasting(tag: Double): Promise<boolean>;
 
-  chromecastState(tag: Int32): Promise<Object>;
+  chromecastState(tag: Double): Promise<Object>;
 
-  airplayState(tag: Int32): Promise<Object>;
+  airplayState(tag: Double): Promise<Object>;
 
-  chromecastStart(tag: Int32): void;
+  chromecastStart(tag: Double): void;
 
-  chromecastStop(tag: Int32): void;
+  chromecastStop(tag: Double): void;
 
-  chromecastJoin(tag: Int32): void;
+  chromecastJoin(tag: Double): void;
 
-  chromecastLeave(tag: Int32): void;
+  chromecastLeave(tag: Double): void;
 
-  airplayStart(tag: Int32): void;
+  airplayStart(tag: Double): void;
 
-  airplayStop(tag: Int32): void;
+  airplayStop(tag: Double): void;
 }
 
 // Note: codegen does not like `TurboModuleRegistry?.getEnforcing`
